sfd
sdfs

+++ src/client/model/ItemCalculations.java
@@ -1,7 +1,7 @@
 package client.model;
 
 public class ItemCalculations {
-	public static boolean isNewBidHigher(int offer, ObservableItem item) {
+	public static boolean isNewBidHigher(double offer, ObservableItem item) {
 		return offer > item.getOfferAmount();
 	}
 }
