+++ src/server/model/item/SaleStrategy/BuyoutStrategy.java
@@ -8,7 +8,7 @@ import java.rmi.RemoteException;
 import java.time.temporal.Temporal;
 
 public class BuyoutStrategy implements SaleStrategy {
-	private int price;
+	private double price;
 	private String buyer;
 
 	public BuyoutStrategy(int price) {
@@ -16,7 +16,7 @@ public class BuyoutStrategy implements SaleStrategy {
 	}
 
 	@Override
-	public void offer(Item item, int amount, String username) {
+	public void offer(Item item, double amount, String username) {
 		price = amount;
 		buyer = username;
 
@@ -34,7 +34,7 @@ public class BuyoutStrategy implements SaleStrategy {
 	}
 
 	@Override
-	public int getOfferAmount() {
+	public double getOfferAmount() {
 		return price;
 	}
